feat: add support for static files
