+++ crates/core/src/lib.rs
@@ -18,7 +18,7 @@ pub mod prelude {
             Body, JsonResponse, JsonResult, NgynContext, NgynRequest, NgynResponse, Param, Query,
             ToBytes, Transducer,
         },
-        traits::{NgynGate, NgynMiddleware},
+        NgynGate, NgynMiddleware,
     };
 }
